sezione16

+++ Models/Entities/Lesson.cs
@@ -5,8 +5,8 @@ namespace MyCourse.Models.Entities
 {
     public partial class Lesson
     {
-        public long Id { get; set; }
-        public long CourseId { get; set; }
+        public int Id { get; set; }
+        public int CourseId { get; set; }
         public string Title { get; set; }
         public string? Description { get; set; }
         public TimeSpan Duration { get; set; }
